Move ControlTabOrderComparer to a separate file
